pyproject.toml
[tool.poetry]
name = "whatabook"
version = "0.1.0"
description = ""
authors = ["EuAlmeida <jvda.vitor.almeida@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
Pillow = "^9.2.0"
django-cpf = "^0.1.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
